Rename c_send to csend for consistency
