+++ lib/rubocop/cop/internal_affairs.rb
@@ -18,7 +18,7 @@ require_relative 'internal_affairs/node_matcher_directive'
 require_relative 'internal_affairs/node_type_predicate'
 require_relative 'internal_affairs/numblock_handler'
 require_relative 'internal_affairs/offense_location_keyword'
-require_relative 'internal_affairs/on_send_without_on_c_send'
+require_relative 'internal_affairs/on_send_without_on_csend'
 require_relative 'internal_affairs/operator_keyword'
 require_relative 'internal_affairs/processed_source_buffer_name'
 require_relative 'internal_affairs/redundant_context_config_parameter'
